Editor lights (first sign of life on PBR material)

+++ src/registry/shader_data_registry/shader_data_basic_lit.go
@@ -54,10 +54,10 @@ func init() {
 type ShaderDataBasicLit struct {
 	rendering.ShaderDataBase
 	Color  matrix.Color
-	Light0 float32
-	Light1 float32
-	Light2 float32
-	Light3 float32
+	Light0 int32
+	Light1 int32
+	Light2 int32
+	Light3 int32
 }
 
 func (t ShaderDataBasicLit) Size() int {
